0924 베타
최종 디자인 컨펌,
데이터베이스 창 내에서 장비 변경 가능
순환펌프, 디비저장주기 등 설정창 기능 완료
경보처리 완료

+++ monitor/templatetags/monitor_extras.py
@@ -22,7 +22,7 @@ def multiply( value, arg ):
     '''
     try:
         value = int( value )
-        arg = float( arg )
+        arg = int( arg )
         if arg: return value * arg
     except: pass
     return ''
